+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [1.1.1] - 2026-06-15
 
 ### Added
-- "Collapse All" button in the Task Explorer panel ([#6](https://github.com/papio-dev/task-cockpit/issues/6))
+- "Collapse All"/"Expand All" button in the Task Explorer panel ([#6](https://github.com/papio-dev/task-cockpit/issues/6))
 - "Find Task" button — quick search and filtering in the task tree (`list.find`)
 
 ## [1.1.0] - 2025-03-20